Introduce helper function for loading LowTCA
In #117, I changed back the hardcoded `loadzlq` instructions
added by 1702bae to `emitLdPackedPtr` as it was before.
That was wrong because LowTCA will be 64 bits on PIE builds
but 32 bits otherwise, whereas `emitLdPackedPtr` uses the size
of `PackedPtr` which is always 64 bits in non-lowptr builds.
Pandemonium ensues as a result.

So, introduce a new helper here that uses the correct size of LowTCA.